Fixed Linux (?) issues
Not sure why these occurred, but on Linux there was a casting error.

+++ Among-Us-Editor/src/club/koupah/aue/gui/types/impl/custom/hostsettings/Float32Setting.java
@@ -25,7 +25,7 @@ public class Float32Setting extends HostSetting {
 	@Override
 	public void update() {
 		if (super.shouldUpdate())
-			((JSpinner) component).setValue(getFloat());
+			((JSpinner) component).setValue(((Float)getFloat()).doubleValue());
 	}
 
 }
